retry libft

+++ ft_lstsize.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_lstsize.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttachi <ttachi@student.42tokyo.jp>         +#+  +:+       +#+        */
+/*   By: ttachi <ttachi@student.42tokyo.ja>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 21:51:07 by ttachi            #+#    #+#             */
-/*   Updated: 2022/10/27 18:11:41 by ttachi           ###   ########.fr       */
+/*   Updated: 2022/11/17 19:42:53 by ttachi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	size_t	lstsize;
+	int	lstsize;
 
 	if (lst == NULL)
 		return (0);
